reorganise graph elements class
